Move paste logic to manual-*-editor
